+++ backend/telekom-analiz-sistemi/telekom/src/main/java/com/pia/telekom/dto/CustomerRequest.java
@@ -20,7 +20,7 @@ public record CustomerRequest(
         String phone,
 
         @NotNull(message = "Bölge seçimi zorunludur")
-        Long regionId,
+        Integer regionId,
 
         String ageGroup,
         String paymentChannelPreference,
