+++ backend/telekom-analiz-sistemi/telekom/src/main/java/com/pia/telekom/dto/InvoiceRequest.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record InvoiceRequest(
-        Long productId,
+        Integer productId,
 
         String paymentChannel,
