+++ backend/telekom-analiz-sistemi/telekom/src/main/java/com/pia/telekom/dto/ProductResponse.java
@@ -3,7 +3,7 @@ package com.pia.telekom.dto;
 import java.math.BigDecimal;
 
 public record ProductResponse(
-        Long productId,
+        Integer productId,
         String name,
         String category,
         BigDecimal monthlyFee,
