+++ backend/telekom-analiz-sistemi/telekom/src/main/java/com/pia/telekom/dto/RechargeResponse.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record RechargeResponse(
-        Long rechargeId, Long customerId, String rechargeChannel,
+        Integer rechargeId, Integer customerId, String rechargeChannel,
         BigDecimal rechargeAmount, LocalDate rechargeDate
 ) {
 }
