+++ backend/telekom-analiz-sistemi/telekom/src/main/java/com/pia/telekom/dto/RegionResponse.java
@@ -3,7 +3,7 @@ package com.pia.telekom.dto;
 import java.math.BigDecimal;
 
 public record RegionResponse(
-        Long regionId,
+        Integer regionId,
         String name,
         String cityType,
         BigDecimal populationWeight
